Update Router, cover with tests. Reorganize tests to different files
